+++ src/components/field-types/TypedFieldComponent.vue
@@ -13,7 +13,8 @@
         :aria-describedby="field.id + '-description'"
         :required="isRequired"
         :disabled="field.disabled"
-        :step="stepSize">
+        :step="stepSize"
+        v-on:keyup="onKeyUp">
 
       <small :id="field.id + '-description'" class="form-text text-muted">
         {{ field.description }}
@@ -27,7 +28,7 @@
 
 <script>
   import VueForm from 'vue-form'
-  import { FormField } from '../../flow.types'
+  import {FormField} from '../../flow.types'
   import FormFieldMessages from '../FormFieldMessages'
 
   const MIN_JAVA_INT = -2147483648
@@ -89,6 +90,12 @@
     methods: {
       toNumber (input) {
         return input !== '' ? Number(input) : null
+      },
+      onKeyUp (event) {
+        // In case of numeric check validity, if invalid place back the old value
+        if (this.isNumberField && event.target.validity && event.target.validity.badInput) {
+          this.localValue = this.value
+        }
       }
     },
     computed: {